Add files via upload
In the classical vehicle routing problem (VRP) the goal is to find optimal routes for vehicles leaving from a depot to serve a specified number of customers. Each customer must be visited once and only once, by exactly one vehicle. Each vehicle must leave from the depot and return there after serving all customers on its route. The objective of this type of problem is are to either minimise the number of vehicles or minimize the fuel cost.
This report focuses on finding the optimal sequence of j trips of n gifts, that must be delivered to several destinations.
For confidentiality purposes, our client X is represented by Santa in this problem and their vehicles; which is the focus of this problem, by reindeers. Our goal is to deliver all the gifts (items) produced, while attempting to minimise the reindeers’ weariness.
To achieve this, we implement two different Stochastic Local Search (SLS) algorithms; Random Search and Simulated Annealing (SA).